Reformat ElasticSearch Dockerfile.
TODO: figure out how to do SSL for the S3 bucket
that ElasticSearch distributes binaries out of.
The issue here is that we access it via a DNS
name that isn’t on the SSL cert, so we have to
use Subject Alternate Name (or maybe even relaxed
host cert verification).