Allow ZK & ES data volumes to succeed on error.
This is because we’re getting all false positives
on error reporting today, owing to the error when
attempting to start an additional container of the
same name as an existing container. 

As this is simply a data volume container, it
shouldn’t fail. Famous last words, but let’s see
how it goes.